NotMyApp::SomeClass may use MooseX::StrictConstructor
Catalyst::Model::Adaptor did send some extra args to the constructor.
Documentation says this should not be the case. A Model using
MooseX::StrictConstructor will die due to this.

_handle_catalystmodel_args is added to lib/Catalyst/Model/Adaptor/Base.pm
to delete these extra args, injected by the Catalyst Model.

t/live-test.t is simply doubled and uses a Model with
MooseX::StrictConstructor. This is done with t/strict-live-test.t and
t/lib/StrictTestApp.pm.

t/strict-live-test.t will only run if MooseX::StrictConstructor is
installed.

Not nice is, that argumentes to be removed are hardcoded in
_handle_catalystmodel_args. But I don't know any other way to get those
dynamically.